Day 6
With two days delay, the solution to Day 6 (mostly). Part 1 returns as the first two places coordinates which have infinite regions, I don't know why they don't get removed by the algorithm. The third and later places in the list are not infinite, the first of those is the correct answer.